DBDE_Rel/requirement.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
django-guardian==1.4.9
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.7
Markdown==3.0.1
Pillow==5.4.1
pycparser==2.18
PyJWT==1.7.1
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
virtualenv==16.0.0
